% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_diagnostics}
\alias{check_diagnostics}
\title{Check diagnostics from a fitted model}
\usage{
check_diagnostics(fit, warn_for_diagnostic_failure = TRUE)
}
\arguments{
\item{fit}{A fitted \code{RtGam} model object}

\item{warn_for_diagnostic_failure}{A bool, whether to issue warnings for
potential diagnostic issues.}
}
\value{
Invisibly, a list of diagnostics. This diagnostic list is also
present in the model object under \code{diagnostics}.
}
\description{
Check diagnostics from a fitted model
}
\examples{
withr::with_seed(12345, {
  cases <- rpois(20, 10)
})
reference_date <- seq.Date(
  from = as.Date("2023-01-01"),
  length.out = 20,
  by = "day"
)
fit <- RtGam::RtGam(cases, reference_date)
check_diagnostics(fit)
}
