% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_elements_non_neg}
\alias{check_elements_non_neg}
\title{Check all elements are 0 or positive}
\usage{
check_elements_non_neg(x, arg = "x", call = rlang::caller_env())
}
\description{
Check all elements are 0 or positive
}
