% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creator.R
\name{dates_to_timesteps}
\alias{dates_to_timesteps}
\title{Convert an arbitrary vector of dates to a vector of timesteps}
\usage{
dates_to_timesteps(reference_date, min_supplied_date, max_supplied_date)
}
\arguments{
\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{min_supplied_date}{The minimum date supplied in \code{reference_date} to the
class constructor.}

\item{max_supplied_date}{The maximum date supplied in \code{reference_date} to the
class constructor.}
}
\value{
A vector of timesteps, corresponding to the supplied reference dates
}
\description{
The \verb{*_supplied_date} arguments are required rather than calculated
internally so that the function can be re-used in methods with a different
\code{reference_date} vector but the same min and max dates. This re-use ensures
that the timestep will match internally but the user can just supply dates
without knowledge of the internal timestep representation.
}
\details{
Timesteps are scaled to be on [0, 1]. This small range is intentional --
I've seen it help numerical samplers converge and by default in R
current-ish dates are large integers.
}
