% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtGam.R
\name{fit}
\alias{fit}
\title{Stub: Will become fitting/class constructor}
\usage{
fit(cases, reference_date, group = NULL)
}
\arguments{
\item{group}{}
}
\value{
Stub function: NULL
}
\description{
Including as a stub function to showcase control flow
}
\examples{
cases <- c(1, 2, 3)
reference_date <- as.Date(c("2023-01-01", "2023-01-02", "2023-01-03"))
mod <- RtGam::fit(cases, reference_date)
}
