% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model.default}
\alias{fit_model.default}
\title{Used to throw informative error if non-supported backend supplied}
\usage{
\method{fit_model}{default}(data, formula, ...)
}
\description{
Used to throw informative error if non-supported backend supplied
}
