% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format_for_return}
\alias{format_for_return}
\title{Format the RtGam object for return from the main function/constructor}
\usage{
format_for_return(fit, df, group, k, m, backend, formula)
}
\arguments{
\item{fit}{The model fit created by \link{fit_model}}

\item{df}{The dataset created by \link{dataset_creator}}
}
\value{
An object of type RtGam
}
\description{
Format the RtGam object for return from the main function/constructor
}
