% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{parse_predict_dates}
\alias{parse_predict_dates}
\title{Convert from user specification to necessary date range}
\usage{
parse_predict_dates(
  object,
  min_date = NULL,
  max_date = NULL,
  horizon = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{object}{An \code{RtGam} object created using the \code{\link[=RtGam]{RtGam()}}
function.}

\item{min_date, max_date}{Optional. Date-like objects specifying the start
and end of the prediction range. See \strong{Details} for more information on
their usage.}

\item{horizon}{Optional. An integer indicating the number of days to
forecast beyond the last date in the model fit. For example, \code{horizon = 7}
predicts the next 7 days.}

\item{call}{The calling environment to be reflected in the error message}
}
\value{
List with two elements: min_date and max_date
}
\description{
Convert from user specification to necessary date range
}
\keyword{internal}
