% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.RtGam}
\alias{plot.RtGam}
\title{Plot method for RtGam fits}
\usage{
\method{plot}{RtGam}(x, parameter = "obs_cases", alpha = 0.05, ...)
}
\arguments{
\item{x}{An \code{RtGam} object created using the \code{\link[=RtGam]{RtGam()}}
function.}

\item{parameter}{A character string specifying the prediction target.
Options are \code{"obs_cases"} (observed cases), \code{"r"} (growth rate), or
\code{"Rt"} (reproduction number). Default is \code{"obs_cases"}.}

\item{alpha}{The opacity of the lines on the plot. Should be between 0 and 1,
with values closer to 0 producing more transparent lines.}

\item{...}{Additional options to pass to \code{\link[=predict.RtGam]{predict.RtGam()}}}
}
\value{
A ggplot2 object
}
\description{
Generates predicted draws from \link{predict.RtGam} and plots them with ggplot2.
See the documentation on \link{predict.RtGam} for options on how to specify
draws.
}
\examples{
fit <- RtGam(
  stochastic_sir_rt[["obs_incidence"]],
  stochastic_sir_rt[["date"]]
)
# Plot draws from the fitted model against the data
plot(fit)

# Plot a 7-day forecast
plot(fit, horizon = 7)

# Plot Rt estimates
plot(fit,
  parameter = "Rt",
  mean_delay = 0,
  gi_pmf = sir_gt_pmf
)
}
