% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_obs_cases}
\alias{predict_obs_cases}
\title{Posterior predicted cases}
\usage{
predict_obs_cases(
  object,
  horizon = NULL,
  min_date = NULL,
  max_date = NULL,
  n = 10,
  gi_pmf = NULL,
  seed = 12345,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{object}{An \code{RtGam} object from \code{\link[=RtGam]{RtGam()}}.}

\item{horizon}{Optional. Integer specifying forecast days from the last date
in the fit. For example, \code{horizon = 7} returns a 7-day forecast.}

\item{min_date}{Optional. A date-like object marking the start of the
prediction period.}

\item{max_date}{Optional. A date-like object marking the end of the
prediction period.}

\item{n}{Number of posterior samples to use. Default is 10.}

\item{gi_pmf}{Optional. A vector representing the generation interval PMF.}

\item{seed}{Random seed for reproducibility. Default is 12345.}

\item{...}{Additional arguments passed to lower-level functions.}
}
\value{
A dataframe with schema ...
}
\description{
TODO put roxygen2 grouping here
}
\keyword{internal}
