% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictor}
\alias{predictor}
\alias{predict_obs_cases}
\alias{predict_growth_rate}
\alias{predict_rt}
\title{Parameter-specific posterior predictions}
\usage{
predict_obs_cases(
  object,
  horizon = NULL,
  min_date = NULL,
  max_date = NULL,
  day_of_week = TRUE,
  n = 10,
  gi_pmf = NULL,
  seed = 12345,
  call = rlang::caller_env(),
  ...
)

predict_growth_rate(
  object,
  horizon = NULL,
  min_date = NULL,
  max_date = NULL,
  n = 10,
  gi_pmf = NULL,
  seed = 12345,
  mean_delay,
  call = rlang::caller_env(),
  ...
)

predict_rt(
  object,
  horizon = NULL,
  min_date = NULL,
  max_date = NULL,
  n = 10,
  gi_pmf = NULL,
  seed = 12345,
  mean_delay,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{object}{An \code{RtGam} object created using the \code{\link[=RtGam]{RtGam()}}
function.}

\item{horizon}{Optional. An integer indicating the number of days to
forecast beyond the last date in the model fit. For example, \code{horizon = 7}
predicts the next 7 days.}

\item{min_date, max_date}{Optional. Date-like objects specifying the start
and end of the prediction range. See \strong{Details} for more information on
their usage.}

\item{day_of_week}{How to handle day-of-week effects when predicting
\code{obs_cases}. Defaults to \code{TRUE}, which identifies and applies the fitted
day-of-week effect, if possible. When automatic detection fails or
different levels are desirable, custom levels can be applied with a vector
of equal length to the number of desired dates. If \code{FALSE}, the day-of-week
effect is turned off (i.e., set to zero). When predicting parameters other
than \code{obs_cases} or \code{object} is an \code{RtGam} model that did not include
day-of-week effects, the day-of-week effect is turned off and this argument
is silently ignored.}

\item{n}{An integer specifying the number of posterior samples to use
for predictions. Default is 100.}

\item{gi_pmf}{Optional. A numeric vector specifying the generation interval
probability mass function (PMF), required when \code{parameter = "Rt"}. The PMF
must be a proper probability distribution (summing to one) with the first
element set to zero to exclude same-day transmission, as required by the
renewal equation. For more information and tools to handle delay
distributions, see the
\href{https://CRAN.R-project.org/package=primarycensored}{primarycensored}
package.}

\item{seed}{An integer specifying the random seed for reproducibility.
Default is 12345.}

\item{call}{The calling environment to pass to error messages}

\item{...}{Additional arguments passed to the underlying sampling functions:
\itemize{
\item When \code{parameter = "obs_cases"}, arguments are passed to \code{gratia: posterior_samples}.
\item When \code{parameter = "r"} or \code{"Rt"}, arguments are passed to \code{gratia: fitted_samples}.
}}

\item{mean_delay}{Optional. An integer specifying the mean number of days
between an individual becoming infected and their case being observed
(e.g., through an emergency department visit or hospitalization). This
value shifts the predictions to account for reporting delays. It is
required when predicting \code{"r"} (growth rate) or \code{"Rt"} (reproduction
number).}
}
\value{
A data frame in
\href{https://www.jstatsoft.org/article/view/v059i10}{tidy format},
where each row represents a posterior draw for a specific date, with the
following columns:
\itemize{
\item \code{reference_date}: The prediction date.
\item \code{.response}: The predicted value for the target parameter.
\item \code{.draw}: The index of the posterior draw.
}

Example output:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  reference_date .response .draw
1     2023-01-01        18     1
2     2023-01-02        13     1
3     2023-01-03        21     1
4     2023-01-01        11     2
5     2023-01-02        19     2
6     2023-01-03        24     2
}\if{html}{\out{</div>}}
}
\description{
Generate posterior draws from an \code{RtGam} fit. Prediction dates can be
specified flexibly using various approaches, and predictions can be drawn for
different model parameters.
}
\details{
Prediction dates can be defined in four ways:
\enumerate{
\item \strong{Default Date Range}: By using only the fit object, predictions are made
across the full date range in the original model.
\item \strong{Using \code{horizon}}: Extends predictions up to \code{horizon} days beyond the
last date in the model fit.
\item \strong{Using \code{min_date} and \code{horizon}}: Predictions start from \code{min_date} and
extend up to \code{horizon} days after the fit’s last date.
\item \strong{Using \code{min_date} and \code{max_date}}: Generates predictions for all dates
within this specified range, inclusive.
}

The \code{mean_delay} parameter adjusts predictions for the
temporal lag between infection and case observation. For example, if
\code{mean_delay = 5},
the model assumes that observed cases reflect infections that occurred on
average five days
earlier. This adjustment ensures that estimates of growth rates (\code{"r"}) and
reproduction
numbers (\code{"Rt"}) align with the correct underlying temporal dynamics.

The \code{parameter} argument determines the type of predictions:
\itemize{
\item \code{"obs_cases"}: Observed cases, including uncertainty from the model's
fit.
\item \code{"r"}: Growth rate, calculated using the centered difference between time
steps.
\item \code{"Rt"}: Reproduction number, incorporating delay distributions and
convolution.
}

Samples are drawn from the posterior distribution of the fitted model using
the \code{gratia} package. The model estimates basis function coefficients on the
smooth terms \eqn{\widehat{\beta}} and smoothing parameter(s) \eqn{\lambda}. The
coefficients have the joint posterior distribution
\deqn{\beta | \lambda \sim N(\widehat{\beta}, \mathbf{V}_{\widehat{\beta}})}
where \eqn{\mathbf{V}_{\widehat{\beta}}} is the smoothing-parameter uncertainty
corrected covariance matrix of the basis function coefficients. We draw
samples from \eqn{\mathbf{V}_{\widehat{\beta}}} and multiply them by the dates of
interest to generate posterior draws. If estimating \code{"Rt"} or \code{"r"}
the day-of-week effect is excluded (i.e., set to zero).

For the intrinsic growth rate, we draw one day before and one day after every
day of interest. We difference these two days within the smooth to get growth
in
the period and divide by two to generate the discrete centered derivative.

For the Rt we map the estimated values on the linear predictor scale back to
the response scale and shift the estimated cases by the mean delay to get
the estimated incident infections (\eqn{I}). We use the incident infections
and the generation interval probability mass function (\eqn{w}) to estimate
Rt via the Cori method: \eqn{I_t / \sum_{s = 1}^{t} I_{t - s} w_s}

For observed incident cases, we apply the estimated negative binomial
observation error to the posterior expected incident cases to generate
posterior predicted incident cases.
}
\references{
Miller, David L. "Bayesian views of generalized additive
modelling." arXiv preprint arXiv:1902.01330 (2019). \cr \cr
Gostic, Katelyn M., et al. "Practical considerations for measuring the
effective reproductive number, Rt." PLoS computational biology 16.12
(2020): e1008409. \cr \cr
Simpson, Gavin L. "Gratia: An R package for exploring generalized additive
models." arXiv preprint arXiv:2406.19082 (2024). \cr \cr
Cori A, Ferguson NM, Fraser C, Cauchemez S. A New Framework and Software to
Estimate Time-Varying Reproduction Numbers During Epidemics. Am J Epidemiol.
2013;178(9):1505–12. pmid:24043437 \cr \cr
}
\seealso{
\code{\link[gratia:fitted_samples]{gratia::fitted_samples()}}, \code{\link[gratia:posterior_samples]{gratia::posterior_samples()}}
}
\keyword{internal}
