% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prep_timesteps_for_pred}
\alias{prep_timesteps_for_pred}
\title{Convert from user-specified dates to internal timesteps}
\usage{
prep_timesteps_for_pred(
  parameter,
  fit_min_date,
  fit_max_date,
  desired_min_date,
  desired_max_date,
  mean_delay,
  call = rlang::caller_env()
)
}
\arguments{
\item{parameter}{A string specifying the prediction parameter. Options are
\code{"obs_cases"}, \code{"incidence"}, \code{"r"}, and \code{"Rt"}. Currently, only
\code{"obs_cases"} is supported. Matching is enforced by \code{\link[rlang:arg_match]{rlang::arg_match()}}.}

\item{mean_delay}{Optional. Numeric mean delay used in the prediction.}
}
\value{
Double vector, the timesteps to predict
}
\description{
Convert from user-specified dates to internal timesteps
}
