% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Parse input vectors into a format for \code{{mgcv}}}
\usage{
prepare_inputs(cases, reference_date, group)
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The grouping variable for the case/reference-date pair. Not yet
implemented and a value other than \code{NULL} will throw an error.}
}
\value{
A dataframe for mgcv
}
\description{
Parse input vectors into a format for \code{{mgcv}}
}
