% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtGam-package.R
\docType{package}
\name{RtGam-package}
\alias{RtGam-package}
\title{RtGam: Estimating Rt with Generalized Additive Models}
\description{
An opinionated implementation of a generalized additive modeling (GAMs) approach to estimate epidemic dynamics. This flexible semi-parametric method allows fitting to arbitrary incidence timeseries, with a natural Bayesian interpretation. Underlying inference methods are implemented via the `{mgcv}` package, which is optimized to produce quick, robust model fits and has well-developed diagnostic methods. These methods are employed here with a focus on _real-time_ use cases. The package is meant for drop-in use in an emerging outbreak, complementing MCMC-based inference methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdcgov/cfa-gam-rt}
  \item \url{https://cdcgov.github.io/cfa-gam-rt/}
  \item Report bugs at \url{https://github.com/cdcgov/cfa-gam-rt/issues}
}

}
\author{
\strong{Maintainer}: Zachary Susswein \email{utb2@cdc.gov} (\href{https://orcid.org/0000-0002-4329-4833}{ORCID})

Authors:
\itemize{
  \item Katelyn Gostic \email{uep6@cdc.gov} (\href{https://orcid.org/0000-0002-9369-6371}{ORCID})
  \item Sam Abbott \email{azw1@cdc.gov} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
}

}
\keyword{internal}
