% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtGam.R
\name{RtGam}
\alias{RtGam}
\title{Fit a generalized additive model to incident cases}
\usage{
RtGam(
  cases,
  reference_date,
  group = NULL,
  k = smooth_dim_heuristic(length(cases))
)
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The geographic grouping for the case/reference-date pair. Not
yet implemented and a value other than \code{NULL} will throw an error.}

\item{k}{An integer, the \emph{total} dimension of all the smoothing basis
functions. Defaults to \code{smooth_dim_heuristic(length(cases))}, which
picks a reasonable estimate based on the number of provided data points.
This total dimension is partitioned between the different smooths in the
model. In the case of diagnostic issues in a fitted \code{RtGam} model,
increasing the value of \code{k} above this default and refitting the model is a
good first step. See the \code{\link[=smooth_dim_heuristic]{smooth_dim_heuristic()}} documentation for
more information.}

\item{m}{An integer, the \emph{total} dimension of the penalty basis for the}
}
\value{
Stub function: NULL
}
\description{
Incident cases are modeled as a smooth function of time with a generalized
additive model (GAM).
}
\details{
If more than three weeks of data are available, \link{RtGam} will fit a GAM with
an adaptive spline basis. This basis is so named because it allows the
wiggliness penalization to vary over time. If one part of the timeseries has
a sudden change in trend while another part shows a smooth increase, the
model can fit both components without smoothing away sharp changes or
introducing additional artificial wiggliness.

The model introduces an additional penalty basis dimension for each
additional 21 days of observed data. A timeseries of 20 or fewer days would
have the same penalty the whole period, a timeseries of 21 to 42 days would
smoothly interpolate between two penalties, and so on for each additional 21
day period. This adaptive penalty increases the computational cost of the
model, but allows for a single model to adapt to changing epidemic dynamics.

In the special case of 20 or fewer observed days, the model will use a single
penalty over the whole period and use a thin-plate spline as the smoothing
basis. The adaptive spline can only use a P-spline smoothing basis. The thin
plate spline generally has better performance and so \link{RtGam} uses it in this
special single-penalty case.
}
\examples{
cases <- c(1, 2, 3)
reference_date <- as.Date(c("2023-01-01", "2023-01-02", "2023-01-03"))
mod <- RtGam::RtGam(cases, reference_date)
}
\seealso{
\code{\link[=smooth_dim_heuristic]{smooth_dim_heuristic()}} more information on the smoothing basis
dimension and \link[mgcv:choose.k]{mgcv::choose.k} for more general guidance on GAMs from
\code{mgcv}
}
