% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sir_gt_pmf}
\alias{sir_gt_pmf}
\title{Generation interval corresponding to the sample \code{stochastic_sir_rt} dataset}
\format{
\code{sir_gt_pmf} A numeric vector of length 26 that sums to one within
numerical tolerance
}
\usage{
sir_gt_pmf
}
\description{
Gostic et al., 2020 simulates data from a stochastic SEIR model. Residence
time in both the E and the I compartments is exponentially distributed, with
a mean of 4 days (or a rate/inverse-scale of 1/4). These residence times
imply a gamma-distributed generation time distribution with a shape of 2 and
a rate of 1/4. The distribution can be regenerated in
\code{data-raw/sir_gt_pmf.R}.
}
\details{
From this parametric specification, we produce a double-censored,
left-truncated probability mass function of the generation interval
distribution. We produce the PMF using
\code{\link[primarycensored:dprimarycensored]{primarycensored::dpcens()}} with version 0.4.0. See
https://doi.org/10.1101/2024.01.12.24301247 for more information on
double-censoring biases and corrections.
}
\references{
Gostic, Katelyn M., et al. "Practical considerations for measuring the
effective reproductive number, Rt." PLoS computational biology 16.12
(2020): e1008409.
}
\keyword{datasets}
