% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_and_renormalize}
\alias{truncate_and_renormalize}
\title{Left-truncate and renormalize to a PMF}
\usage{
truncate_and_renormalize(pmf, tol = 1e-10)
}
\arguments{
\item{pmf}{A probability mass function formatted as a vector. Usually a
generation interval indexed from day 0.}

\item{tol}{The numerical tolerance applied to the sum-to-one check on the
renormalized PMF.}
}
\value{
A probability mass function with no density in the first bin.
}
\description{
The renewal equation assumes no same-day transmission. This assumption
means that when working with a generation interval formatted as a PMF
we need to remove the element corresponding to day zero. This function
drops the element and renormalizes the remaining density to sum to one.
}
