% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{type_checker}
\alias{type_checker}
\alias{check_date}
\alias{check_vector}
\alias{check_integer}
\title{Basic type checks for input validation}
\usage{
check_date(x, arg = "x", call = rlang::caller_env())

check_vector(x, arg = "x", call = rlang::caller_env())

check_integer(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{Object with type checking}

\item{arg}{Name of the argument supplying the object}

\item{call}{Calling environment to be passed to the type checker}
}
\value{
NULL, invisibly
}
\description{
Basic type checks for input validation
}
