% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_dates}
\alias{validate_dates}
\title{Can be re-used for both reference and report dates}
\usage{
validate_dates(dates, arg, call)
}
\description{
Can be re-used for both reference and report dates
}
