% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_diagnostics}
\alias{check_diagnostics}
\title{Check diagnostics from a fitted model}
\usage{
check_diagnostics(fit, warn_for_diagnostic_failure = TRUE)
}
\arguments{
\item{fit}{A fitted \code{RtGam} model object}

\item{warn_for_diagnostic_failure}{A bool, whether to issue warnings for
potential diagnostic issues.}
}
\value{
Invisibly, a list of diagnostics. This diagnostic list is also
present in the model object under \code{diagnostics}.
}
\description{
Check diagnostics from a fitted model
}
\examples{
fit <- RtGam::RtGam(
  cases = c(1L, 2L, 3L, 4L),
  reference_date = as.Date(c(
    "2023-01-01",
    "2023-01-02",
    "2023-01-03"
  ))
)
check_diagnostics(fit)
}
