% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creator.R
\name{dataset_creator}
\alias{dataset_creator}
\title{Parse input vectors into a format for \code{{mgcv}}}
\usage{
dataset_creator(cases, reference_date, group, backend)
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The grouping variable for the case/reference-date pair. Not yet
implemented and a value other than \code{NULL} will throw an error.}

\item{backend}{One of \code{gam} or \code{bam}; defaults to \code{gam}. In general, models
should be fit with \code{\link[mgcv:gam]{mgcv::gam()}}. If \code{\link[mgcv:gam]{mgcv::gam()}} is too slow,
\code{\link[mgcv:bam]{mgcv::bam()}} converges more quickly but introduces some additional
numerical error. Note that the \code{bam} backend uses the \code{discrete = TRUE}
option for an additional speedup. See \code{\link[mgcv:bam]{mgcv::bam()}} for more information.}
}
\value{
A dataframe for mgcv
}
\description{
Parse input vectors into a format for \code{{mgcv}}
}
