% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{parse_predict_dates}
\alias{parse_predict_dates}
\title{Convert from user specification to necessary date range}
\usage{
parse_predict_dates(
  object,
  min_date = NULL,
  max_date = NULL,
  horizon = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{object}{An \code{RtGam} object from \code{\link[=RtGam]{RtGam()}}.}

\item{min_date, max_date}{Optional. A date-like object. See details for more
information on specification.}

\item{horizon}{Optional. Integer specifying forecast days from the last date
in the fit. For example, \code{horizon = 7} returns a 7-day forecast.}

\item{call}{The calling environment to be reflected in the error message}
}
\value{
List with two elements: min_date and max_date
}
\description{
Convert from user specification to necessary date range
}
\keyword{internal}
