% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.RtGam}
\alias{predict.RtGam}
\title{Predict Method for RtGam Models}
\usage{
\method{predict}{RtGam}(
  object,
  parameter = "obs_cases",
  horizon = NULL,
  min_date = NULL,
  max_date = NULL,
  n = 100,
  mean_delay = NULL,
  gi_pmf = NULL,
  seed = 12345,
  ...
)
}
\arguments{
\item{object}{An \code{RtGam} object from \code{\link[=RtGam]{RtGam()}}.}

\item{parameter}{A string specifying the prediction parameter. Options are
\code{"obs_cases"}, \code{"r"}, and \code{"Rt"}. Defaults to \code{"obs_cases"}.}

\item{horizon}{Optional. Integer specifying forecast days from the last date
in the fit. For example, \code{horizon = 7} returns a 7-day forecast.}

\item{min_date, max_date}{Optional. A date-like object. See details for more
information on specification.}

\item{n}{Number of posterior samples to use. Default is 100.}

\item{mean_delay}{Optional. Numeric mean delay used in the prediction.}

\item{gi_pmf}{Optional. A vector representing the generation interval PMF.}

\item{seed}{Random seed for reproducibility. Default is 12345.}

\item{...}{Additional arguments passed to lower-level functions.}
}
\value{
A dataframe in \href{https://www.jstatsoft.org/article/view/v059i10}{tidy format},
with each row representing a draw for a specific date:
\itemize{
\item \code{reference_date}: Date of the prediction.
\item \code{.response}: Predicted value (e.g., observed cases).
\item \code{.draw}: ID of the posterior draw.
}

Example output:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  reference_date .response .draw
1     2023-01-01        18     1
2     2023-01-02        13     1
3     2023-01-03        21     1
4     2023-01-01        11     2
5     2023-01-02        19     2
6     2023-01-03        24     2
}\if{html}{\out{</div>}}
}
\description{
Generates predictions from an \code{RtGam} fit. Prediction dates can be specified
flexibly using multiple approaches.
}
\details{
Prediction dates can be set in four ways:
\enumerate{
\item \strong{Using Fit Object Alone}: Predictions span the full date range in the
original model fit.
\item \strong{Using \code{horizon}}: Forecasts extend \code{horizon} days from the fit’s last
date.
\item \strong{Using \code{min_date} and \code{horizon}}: Predictions start at \code{min_date} and
end \code{horizon} days after the fit’s last date.
\item \strong{Using \code{min_date} and \code{max_date}}: Predictions span all dates between
these two (inclusive).
}
}
