% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.RtGam}
\alias{print.RtGam}
\title{Print an RtGam object}
\usage{
\method{print}{RtGam}(x, ...)
}
\arguments{
\item{x}{Fitted model object of class RtGam}
}
\value{
The RtGam object, invisibly
}
\description{
Print an RtGam object
}
