% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtGam.R
\name{RtGam}
\alias{RtGam}
\title{Stub: Will become fitting/class constructor}
\usage{
RtGam(cases, reference_date, group = NULL)
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The geographic grouping for the case/reference-date pair. Not
yet implemented and a value other than \code{NULL} will throw an error.}
}
\value{
Stub function: NULL
}
\description{
Including as a stub function to showcase control flow
}
\examples{
cases <- c(1, 2, 3)
reference_date <- as.Date(c("2023-01-01", "2023-01-02", "2023-01-03"))
mod <- RtGam::RtGam(cases, reference_date)
}
