% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtGam.R
\name{RtGam}
\alias{RtGam}
\title{Fit a generalized additive model to incident cases}
\usage{
RtGam(
  cases,
  reference_date,
  group = NULL,
  k = smooth_dim_heuristic(length(cases)),
  m = penalty_dim_heuristic(length(unique(reference_date))),
  backend = "gam",
  warn_for_diagnostic_failure = TRUE,
  ...
)
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The grouping variable for the case/reference-date pair. Not yet
implemented and a value other than \code{NULL} will throw an error.}

\item{k}{An integer, the \emph{total} dimension of all the smoothing basis
functions. Defaults to \code{smooth_dim_heuristic(length(cases))}, which picks a
reasonable estimate based on the number of provided data points. This total
dimension is partitioned between the different smooths in the model. In the
case of diagnostic issues in a fitted \code{RtGam} model, increasing the value
of \code{k} above this default and refitting the model is a good first step. See
the \code{\link[=smooth_dim_heuristic]{smooth_dim_heuristic()}} documentation for more information.}

\item{m}{An integer, the dimension of the penalty basis for the global smooth
trend. If \code{m} is greater than 1, the smooth's wiggliness can change over
time. An increase in this value above the default should be done carefully.
See \code{\link[=penalty_dim_heuristic]{penalty_dim_heuristic()}} for more information on \code{m} and when to
consider changing the default.}

\item{backend}{One of \code{gam} or \code{bam}; defaults to \code{gam}. In general, models
should be fit with \code{\link[mgcv:gam]{mgcv::gam()}}. If \code{\link[mgcv:gam]{mgcv::gam()}} is too slow,
\code{\link[mgcv:bam]{mgcv::bam()}} converges more quickly but introduces some additional
numerical error. Note that the \code{bam} backend uses the \code{discrete = TRUE}
option for an additional speedup. See \code{\link[mgcv:bam]{mgcv::bam()}} for more information.}

\item{warn_for_diagnostic_failure}{Should warnings be issued for
automatically identified diagnostic issues? Defaults to true. A list of
quantitative model diagnostics can be inspected in the \code{diagnostics} slot
of the returned \code{RtGam} object.}

\item{...}{Additional arguments passed to the specified modelling backend.
For example, the default negative binomial error structure could be changed
to poisson in the default \link[mgcv:gam]{mgcv::gam} backend by passing \code{family = "poisson"}.}
}
\value{
Stub function: NULL
}
\description{
Incident cases are modeled as a smooth function of time with a generalized
additive model (GAM). The model is fit with \code{\link[mgcv:gam]{mgcv::gam()}} and some
familiarity with \code{mgcv} may be helpful.
}
\section{Model specification}{
Incident cases (\eqn{y}) are modeled as smoothly changing over time:

\deqn{\text{log}\{E(y)\} = \alpha + f_{\text{global}(t)}}

where incidence is negative-binomially distributed and \eqn{f(t)} is a smooth
function of time.
}

\examples{
withr::with_seed(12345, {
  cases <- rpois(20, 10)
})
reference_date <- seq.Date(
  from = as.Date("2023-01-01"),
  length.out = 20,
  by = "day"
)
fit <- RtGam::RtGam(cases, reference_date)
fit
}
\seealso{
\code{\link[=smooth_dim_heuristic]{smooth_dim_heuristic()}} more information on the smoothing basis
dimension, \link[mgcv:choose.k]{mgcv::choose.k} for more general guidance on GAMs from \code{mgcv},
and \link[mgcv:gam]{mgcv::gam}/\link[mgcv:bam]{mgcv::bam} for documentation on arguments to the model
fitting functions.
}
