% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate user-provided inputs for type correctness}
\usage{
validate(cases, reference_date, group, call = rlang::caller_env())
}
\arguments{
\item{cases}{A vector of non-negative incident case counts occurring on an
associated \code{reference_date}. Missing values (NAs) are not allowed.}

\item{reference_date}{The associated date on which the count of incident
\code{cases} occurred. Missing dates are not allowed and dates can only occur
once.}

\item{group}{The geographic grouping for the case/reference-date pair. Not
yet implemented and a value other than \code{NULL} will throw an error.}

\item{call}{Caller environment to be passed through the stack}
}
\description{
Does not validate \emph{sensibility} of inputs nor does it do any
comparisons between the inputs to check lengths
}
